[cleanup] Move verify_stats to utility.py
